feat: Implement a class to manage all cards
Implement the Compendium class to store all cards in the game.
Afterwards, the stored cards are displayed to the game in order
to prove proper loading of the textures. Also, some game config
properties -- such as: canvas width, canvas height, and canvas
background color -- change for a better focus on the fundamental
game systems.